@@ -1,11 +1,8 @@
 <?php
 
-
 namespace App\Entity;
-
 
 class SolrResponse
 {
-    /** @var Response $response */
-    public $response;
+    public SolrDocsResponse $response;
 }
